Moved all classes relating to Robot State Management to their own folder, renamed color to signaling.
