+++ src/main/java/frc/robot/stateManagement/NoteState.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.stateManagement;
 
 import java.util.Arrays;
