+++ src/main/java/frc/robot/stateManagement/RobotStateManager.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.stateManagement;
 
 import edu.wpi.first.networktables.BooleanPublisher;
 import edu.wpi.first.networktables.BooleanTopic;
